refactor controller class
